Move default profile save to utils
